move cppwinrt less exception helper, use in DevSupportManager
